+++ src/main/java/domain/ladder/RandomStickGenerator.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.ladder;
 
 import java.util.Random;
