+++ src/main/java/domain/ladder/Stick.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.ladder;
 
 public enum Stick {
