+++ src/main/java/domain/ladder/StickGenerator.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.ladder;
 
 public interface StickGenerator {
