+++ src/test/java/domain/LineTest.java
@@ -1,5 +1,8 @@
 package domain;
 
+import domain.ladder.Line;
+import domain.ladder.Stick;
+import domain.ladder.StickGenerator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
